refactor: member DTO 클래스 DTO 패키지에서 보관
