+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/dto/GatheringPreview.java
@@ -4,7 +4,7 @@ import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
 import com.triplem.momoim.core.domain.gathering.model.Gathering;
-import com.triplem.momoim.core.domain.member.GatheringMemberDetail;
+import com.triplem.momoim.core.domain.member.dto.GatheringMemberDetail;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
